Fix simd instructions (#167)
* fix SIMD compiling issue in linux

* fix immintrin.h

* fix linux simd instructions compiling issue in gcc-7 with cuda 10.2
