+++ AnnService/CoreLibrary.vcxproj.filters
@@ -228,6 +228,9 @@
     <ClCompile Include="src\Core\Common\InstructionUtils.cpp">
       <Filter>Source Files\Core\Common</Filter>
     </ClCompile>
+    <ClCompile Include="src\Core\Common\DistanceUtils.cpp">
+      <Filter>Source Files\Core\Common</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
